Add a license exception clause for Apache 2.0
rmtools uses the requests package which is licensed under the Apache 2.0
license which some interpret as being incompatible with the GPLv2.  This
exclusion clause clarifies the licensing situation by explicitly
allowing packages with that license to be used with this program. This
does have the limitation that other plain GPLv2 packages cannot be
incorporated in the future, but that is currently not a problem.